+++ src/main/java/pl/javastart/sellegro/auction/AuctionFilters.java
@@ -7,35 +7,4 @@ public class AuctionFilters {
     private String carModel;
     private String color;
 
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getCarMaker() {
-        return carMaker;
-    }
-
-    public void setCarMaker(String carMaker) {
-        this.carMaker = carMaker;
-    }
-
-    public String getCarModel() {
-        return carModel;
-    }
-
-    public void setCarModel(String carModel) {
-        this.carModel = carModel;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public void setColor(String color) {
-        this.color = color;
-    }
 }
