--- src/main/java/pl/javastart/sellegro/home/HomeController.java
@@ -1,22 +1,0 @@
-package pl.javastart.sellegro.home;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import pl.javastart.sellegro.auction.AuctionService;
-
-@Controller
-public class HomeController {
-
-    private AuctionService auctionService;
-
-    public HomeController(AuctionService auctionService) {
-        this.auctionService = auctionService;
-    }
-
-    @GetMapping("/")
-    public String home(Model model) {
-        model.addAttribute("cars", auctionService.find4MostExpensive());
-        return "home";
-    }
-}
