--- src/test/java/pl/javastart/sellegro/SellegroApplicationTests.java
@@ -1,16 +1,0 @@
-package pl.javastart.sellegro;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SellegroApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
